--- app/src/main/java/com/kozyrev/simbirtraineeship/ProfileActivity.java
@@ -1,43 +1,0 @@
-package com.kozyrev.simbirtraineeship;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-import android.view.Menu;
-import android.widget.TextView;
-
-import androidx.appcompat.widget.Toolbar;
-
-import com.google.android.material.bottomnavigation.BottomNavigationView;
-
-public class ProfileActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_profile);
-        setupToolbar();
-        setupBottomNavigationView();
-    }
-
-    private void setupToolbar() {
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        TextView toolbar_title = findViewById(R.id.toolbar_title);
-
-        setSupportActionBar(toolbar);
-        toolbar_title.setText(R.string.nav_profile);
-
-        getSupportActionBar().setDisplayShowTitleEnabled(false);
-    }
-
-    private void setupBottomNavigationView(){
-        BottomNavigationView bottomNavigationView = findViewById(R.id.bottom_navigation);
-        bottomNavigationView.setSelectedItemId(R.id.navigation_profile);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu){
-        getMenuInflater().inflate(R.menu.top_menu_profile, menu);
-        return super.onCreateOptionsMenu(menu);
-    }
-}
