+++ app/src/main/java/com/kozyrev/simbirtraineeship/ProfileFragment.java
@@ -1,0 +1,25 @@
+package com.kozyrev.simbirtraineeship;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+public class ProfileFragment extends Fragment {
+
+    public ProfileFragment() {}
+
+    public static ProfileFragment newInstance(String param1, String param2) {
+        ProfileFragment fragment = new ProfileFragment();
+        Bundle args = new Bundle();
+        fragment.setArguments(args);
+        return fragment;
+    }
+
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.fragment_profile, container, false);
+        return view;
+    }
+}
