+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Tue Nov 19 23:27:56 SAMT 2019
-sdk.dir=C\:\\Users\\Android\\Sdk
+#Wed Nov 20 13:10:26 GMT+04:00 2019
+sdk.dir=C\:\\Users\\kozyrev_aa\\AppData\\Local\\Android\\Sdk
